+++ minimum_dependency_generator/main.py
@@ -1,20 +1,22 @@
 from argparse import ArgumentParser
 
-from minimum_dependency_generator import generate_min_requirements, parse_setupcfg
+from minimum_dependency_generator import generate_min_requirements
 
 
 def main():
     parser = ArgumentParser(description="reads a requirements file and outputs the minimized requirements")
-    parser.add_argument('--requirements_paths', nargs='+',
+
+    parser.add_argument('--paths', nargs='+',
                         help='path for requirements to minimize', required=True)
-    parser.add_argument('--options_names', nargs='+',
+
+    parser.add_argument('--options', nargs='+', default=None,
                         help='path for requirements to minimize')
+
+    parser.add_argument('--extras_require', nargs='+', default=None,
+                        help='path for requirements to minimize')
+
     args = parser.parse_args()
-    if args.options_names:
-        requirements = parse_setupcfg(args.requirements_paths, parse_setupcfg)
-    else:
-        requirements = generate_min_requirements(args.requirements_paths)
-        print(requirements)
+    requirements = generate_min_requirements(args.paths, args.options, args.extras_require)
     requirements = sanitize_string(requirements)
     # DO NOT remove, the GH action needs to output
     print("::set-output name=min_reqs::{}".format(requirements))
