+++ minimum_dependency_generator/tests/conftest.py
@@ -1,0 +1,46 @@
+import pytest
+
+
+@pytest.fixture(scope="session", autouse=True)
+def pandas_dep():
+    return "pandas>=0.24.1,<2.0.0,!=1.1.0,!=1.1.1"
+
+
+@pytest.fixture(scope="session", autouse=True)
+def woodwork_dep():
+    return "woodwork==0.0.11"
+
+
+@pytest.fixture(scope="session", autouse=True)
+def dask_dep():
+    return "dask[dataframe]>=2.30.0,<2012.12.0"
+
+
+@pytest.fixture(scope="session", autouse=True)
+def ploty_dep():
+    return "plotly>=4.14.0"
+
+
+@pytest.fixture(scope="session", autouse=True)
+def numpy_lower():
+    return "numpy>=1.15.4"
+
+
+@pytest.fixture(scope="session", autouse=True)
+def numpy_upper():
+    return "numpy<1.20.0"
+
+
+@pytest.fixture(scope="session", autouse=True)
+def other_req_path():
+    return "-r core-requirements.txt"
+
+
+@pytest.fixture(scope="session", autouse=True)
+def setuptools():
+    return "setuptools >= 47"
+
+
+@pytest.fixture(scope="session", autouse=True)
+def p_ytest_dep():
+    return "pytest >= 5.2.0"
